+++ json-path/src/main/java/com/jayway/jsonpath/spi/http/HttpProvider.java
@@ -18,6 +18,7 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.net.URL;
 
+@Deprecated
 public interface HttpProvider {
 
     InputStream get(URL url) throws IOException;