+++ json-path/src/main/java/com/jayway/jsonpath/spi/json/AbstractJsonProvider.java
@@ -12,11 +12,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.jayway.jsonpath.internal.spi.json;
+package com.jayway.jsonpath.spi.json;
 
 import com.jayway.jsonpath.JsonPathException;
-import com.jayway.jsonpath.spi.json.JsonProvider;
-
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;